payload: move requirements code to a separate file
Split the payload initiative!
